refactor: cahnge middleware
