+++ apps/competitions-api/src/routes/competitions/getCompetitionByEid.ts
@@ -1,22 +1,23 @@
 import { Router } from 'express';
 import { prisma } from '@competition-manager/prisma';
 import { parseRequest } from '@competition-manager/utils';
-import { Competition$ } from '@competition-manager/schemas';
+import { Eid$ } from '@competition-manager/schemas';
+import { z } from 'zod';
 
 export const router = Router();
 
-const Params$ = Competition$.pick({
-    eid: true,
+const Params$ = z.object({
+    competitionEid: Eid$
 });
 
 router.get(
     '/:eid',
     parseRequest('params', Params$),
     async (req, res) => {
-        const { eid } = Params$.parse(req.params);
+        const { competitionEid } = Params$.parse(req.params);
         const competition = await prisma.competition.findUnique({
             where: {
-                eid,
+                eid: competitionEid,
                 publish: true
             },
             select: {
@@ -31,7 +32,6 @@ router.get(
                 oneDayBibStart: true,
                 events: true,
             }
-
         });
         if (!competition) {
             res.status(404).send('Competition not found');
