+++ apps/competitions-api/src/routes/events/createEvent.ts
@@ -1,8 +1,8 @@
 import { Router } from 'express';
 import { prisma } from '@competition-manager/prisma';
 import { z } from 'zod';
-import { parseRequest, authMiddleware, adminAuthMiddleware } from '@competition-manager/utils';
-import { CompetitionEvent$ } from '@competition-manager/schemas';
+import { parseRequest, checkRole, checkAdminRole, AuthenticatedRequest } from '@competition-manager/utils';
+import { AdminFromCompetition, AdminFromCompetiton$, CompetitionEvent$ } from '@competition-manager/schemas';
 
 export const router = Router();
 
@@ -34,13 +34,34 @@ router.post(
     '/:competitionEid/events',
     parseRequest('params', Params$),
     parseRequest('body', Body$),
-    authMiddleware('admin'),
-    adminAuthMiddleware('competitions'),
-    async (req, res) => {
+    checkRole('admin'),
+    async (req: AuthenticatedRequest, res) => {
         const { competitionEid } = Params$.parse(req.params);
         const body = Body$.parse(req.body);
         const competitionEvent = NewCompetitionEvent$.parse(body);
         const { eventId, categoriesId, parentId } = body;
+        const competition = await prisma.competition.findUnique({
+            where: {
+                eid: competitionEid
+            },
+            include: {
+                admins: {
+                    select: {
+                        access: true,
+                        userId: true
+                    }
+                }
+            }
+        });
+        if (!competition) {
+            res.status(404).send('Competition not found');
+            return;
+        }
+        if (!checkAdminRole('competitions', req.user!.id, z.array(AdminFromCompetiton$).parse(competition.admins), res)) {
+            return;
+        }
+
+
         const event = await prisma.event.findUnique({
             where: {
                 id: eventId
