+++ shared-packages/schemas/src/Admin.ts
@@ -11,5 +11,10 @@ export const Admin$ = z.object({
     competitionId: z.number().positive(),
     access: z.array(z.enum(ACCESS)),
 });
-
 export type Admin = z.infer<typeof Admin$>;
+
+export const AdminFromCompetiton$ = Admin$.pick({
+    userId: true,
+    access: true
+});
+export type AdminFromCompetition = z.infer<typeof AdminFromCompetiton$>;
