+++ shared-packages/schemas/src/Eid.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+
+
+export const Eid$ = z.string();
+
+export type Eid = z.infer<typeof Eid$>;
