+++ shared-packages/schemas/src/index.ts
@@ -11,3 +11,4 @@ export * from './Result';
 export * from './TokenData';
 export * from './User';
 export * from './UserPreferences';
+export * from './Eid';
