+++ shared-packages/utils/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./isBestResult";
 export * from "./jwtToken";
 export * from "./middleware/index";
+export * from "./checkAdminRole";
